+++ sdk/include/pixie/pixie16/module.hpp
@@ -718,7 +718,7 @@ protected:
     /*
      * Set the bus device number. Only used when simulating.
      */
-    virtual void set_bus_device_number(int device_number);
+    virtual void set_bus_device_number(size_t device_number);
 
     /*
      * Load the variable address map.
